Keep single parameter hugged when the return type hugs an object type
