+++ src/language-js/print/assignment.js
@@ -12,6 +12,7 @@ import getStringWidth from "../../utilities/get-string-width.js";
 import isNonEmptyArray from "../../utilities/is-non-empty-array.js";
 import { getCallArguments } from "../utilities/call-arguments.js";
 import { CommentCheckFlags, hasComment } from "../utilities/comments.js";
+import { getTypeParametersFromTypeReference } from "../utilities/get-type-parameters-from-type-reference.js";
 import { hasLeadingOwnLineComment } from "../utilities/has-leading-own-line-comment.js";
 import { isIndentableBlockComment } from "../utilities/indentable-block-comment.js";
 import { isLoneShortArgument } from "../utilities/is-lone-short-argument.js";
@@ -340,19 +341,6 @@ function isArrowFunctionVariableDeclarator(node) {
     node.type === "VariableDeclarator" &&
     node.init?.type === "ArrowFunctionExpression"
   );
-}
-
-function getTypeParametersFromTypeReference(node) {
-  let typeArguments;
-  switch (node.type) {
-    case "GenericTypeAnnotation":
-      typeArguments = node.typeParameters;
-      break;
-    case "TSTypeReference":
-      typeArguments = node.typeArguments;
-      break;
-  }
-  return typeArguments?.params;
 }
 
 /**
