+++ src/language-js/utilities/get-type-parameters-from-type-reference.js
@@ -1,0 +1,14 @@
+function getTypeParametersFromTypeReference(node) {
+  let typeArguments;
+  switch (node.type) {
+    case "GenericTypeAnnotation":
+      typeArguments = node.typeParameters;
+      break;
+    case "TSTypeReference":
+      typeArguments = node.typeArguments;
+      break;
+  }
+  return typeArguments?.params;
+}
+
+export { getTypeParametersFromTypeReference };
