+++ tests/format/typescript/hugged-return-type/format.test.js
@@ -1,0 +1,2 @@
+runFormatTest(import.meta, ["typescript"]);
+runFormatTest(import.meta, ["typescript"], { objectWrap: "collapse" });
