+++ tests/format/typescript/hugged-return-type/issue-19968.ts
@@ -1,0 +1,7 @@
+async function case1(row: BingConnectionRow): Promise<{ row: BingConnectionRow; bundle: BingCredentialBundle; accessToken: string; authorizationVersion: number }> {
+  return null as never;
+}
+
+function case2(row: BingConnectionRow): Record<string, { row: BingConnectionRow; bundle: BingCredentialBundle; accessToken: string; authorizationVersion: number }> {
+  return null as never;
+}
